+++ lib/uniform_notifier.rb
@@ -13,16 +13,17 @@ require 'uniform_notifier/rollbar'
 require 'uniform_notifier/bugsnag'
 require 'uniform_notifier/slack'
 require 'uniform_notifier/raise'
+require 'uniform_notifier/terminal_notifier'
 
 class UniformNotifier
   AVAILABLE_NOTIFIERS = [:alert, :console, :growl, :honeybadger, :xmpp, :rails_logger,
                          :customized_logger, :airbrake, :rollbar, :bugsnag, :slack, :raise,
-                         :sentry
+                         :sentry, :terminal_notifier
                         ]
 
   NOTIFIERS = [JavascriptAlert, JavascriptConsole, Growl, HoneybadgerNotifier, Xmpp, RailsLogger,
                CustomizedLogger, AirbrakeNotifier, RollbarNotifier, BugsnagNotifier, Raise, Slack,
-               SentryNotifier
+               SentryNotifier, TerminalNotifier
               ]
 
   class NotificationError < StandardError; end